[FEAT] #26 - ImageListComponent 위치 이동
